Fix AUTOSAR C++14 M5-0-4 violations in DatarouterMessageClientImpl
- Change integer literals 0 and 10 to unsigned literals 0U and 10U in
  IClientFactory::ClientConfig initialization.
- Change static_cast<int> to static_cast<std::ptrdiff_t> in std::advance
  to avoid implicit signedness conversion from unsigned to signed type.

More explaination:
std::advance expects a signed type: The second parameter of std::advance should be the iterator's difference_type, which is a signed integer type. For std::string::iterator, this is typically std::ptrdiff_t.

Original violation: MessagePassingConfig::kRandomFilenameStartIndex is std::size_t (unsigned), and casting it to int creates an implicit signedness conversion, violating AUTOSAR M5-0-4.

std::ptrdiff_t is the standard type for iterator distances: According to the C++ standard, std::ptrdiff_t is specifically designed for:
= Pointer arithmetic results.
= Iterator difference operations.
= Signed versions of size operations.

Better than int:
= std::ptrdiff_t matches the address space size (32-bit on 32-bit systems, 64-bit on 64-bit systems).
= Prevents potential overflow issues with large container sizes.

So std::ptrdiff_t is the semantically correct and standards-compliant type for this operation, avoiding the AUTOSAR violation while ensuring type safety.

refer to: https://en.cppreference.com/w/cpp/types/ptrdiff_t.html

GIT_ORIGIN_SPP_REV_ID: d20d91dd28b6d0fa9a3f8a155119526a0002bf56

+++ score/mw/log/detail/data_router/data_router_message_client_impl.cpp
@@ -272,7 +272,7 @@ void DatarouterMessageClientImpl::SendConnectMessage() noexcept
          (MessagePassingConfig::kRandomFilenameStartIndex + msg.GetRandomPart().size() + static_cast<std::size_t>(1))))
     {
         auto start_iter = writer_file_name_.begin();
-        std::advance(start_iter, static_cast<int>(MessagePassingConfig::kRandomFilenameStartIndex));
+        std::advance(start_iter, static_cast<std::ptrdiff_t>(MessagePassingConfig::kRandomFilenameStartIndex));
         auto random_part = msg.GetRandomPart();
         std::ignore = std::copy_n(start_iter, random_part.size(), random_part.begin());
         msg.SetRandomPart(random_part);
@@ -325,7 +325,7 @@ score::cpp::expected_blank<score::os::Error> DatarouterMessageClientImpl::Create
         MessagePassingConfig::kMaxMessageSize,
         MessagePassingConfig::kMaxReplySize,
         MessagePassingConfig::kMaxNotifySize};
-    const score::message_passing::IClientFactory::ClientConfig client_config{0, 10, false, true, false};
+    const score::message_passing::IClientFactory::ClientConfig client_config{0U, 10U, false, true, false};
     sender_ = message_passing_factory_->CreateClient(protocol_config, client_config);
 
     if (sender_ == nullptr)
